+++ bindings/wasm/tests-ts/css_inline.spec.ts
@@ -9,14 +9,14 @@ describe("CSS inliner", () => {
           "<html><head><title>Test</title><style>h1 { color:red; }</style></head><body><h1>Test</h1></body></html>"
         )
       ).to.equal(
-        '<html><head><title>Test</title><style>h1 { color:red; }</style></head><body><h1 style="color:red;">Test</h1></body></html>'
+        '<html><head><title>Test</title></head><body><h1 style="color:red;">Test</h1></body></html>'
       );
     });
-    it("style tag is removed", function () {
+    it("style tag is kept", function () {
       expect(
         inline(
           "<html><head><title>Test</title><style>h1 { color:red; }</style></head><body><h1>Test</h1></body></html>",
-          { remove_style_tags: false }
+          { keep_style_tags: true }
         )
       ).to.equal(
         '<html><head><title>Test</title><style>h1 { color:red; }</style></head><body><h1 style="color:red;">Test</h1></body></html>'