+++ css-inline/src/html/selectors/attr_value.rs
@@ -3,7 +3,7 @@ use html5ever::tendril::StrTendril;
 use std::{fmt, fmt::Write};
 
 /// To use `StrTendril` in selectors, we need to implement `ToCss` on a wrapper.
-#[derive(Clone, Debug, Default, Eq, Hash, PartialEq)]
+#[derive(Clone, Debug, Eq, PartialEq)]
 pub(crate) struct AttrValue(StrTendril);
 
 impl ToCss for AttrValue {